1) added a pruning function to emiliminate the presence of the UMLS id for "Genes" and the string id "Genes"
2) updated print statements to catch the total number of aliases/ids
for [gene] --> predicate --> gene 1-hop queries.

3) still have the 2-hop query commented out. This is due to
the large number of edges produced by the 2-hop. (ex:

[drug] --> predicate --> [genes from 1 hop] --> predicate --> target-gene

with the KCNN3 example, the number of edges on the 2nd hop is still
in the thousands. I need to build in some custom logic to handle these
edges at the time of generation to filter them by some utility/test,
such as the FDA approval status of the 2-hop drug. More on this later.